fix: old talks were being sent in email notification
A talk is mapped to a booking by external_id field. A booking with
external id talks.1234 is linked to an entry in talks table with id
1234.

When sending email, we fetch upcoming bookings and read the external_id.
Which is split into two fields using explode('.', ...) and only
numerical part was being used to fetch the talk. This is an error since
a NBJC entry e.g. NBJC.342 is now being mapped onto talks.342 (which is
a very old talk) and being sent out as if this belongs to the event with
external id NBJC.342. This has been fixed.